Move stop_service into DockerCompose
