+++ shotover-proxy/Cargo.toml
@@ -93,7 +93,6 @@ metrics-util = "0.14.0"
 cdrs-tokio = { git = "https://github.com/krojew/cdrs-tokio", branch = "8.0-dev" }
 scylla = { version = "0.6.0", features = ["ssl"] }
 rstest = "0.15.0"
-docker-api = "0.11.0"
 
 [[bench]]
 name = "benches"
