+++ test-helpers/Cargo.toml
@@ -14,3 +14,4 @@ anyhow = "1.0.42"
 regex = "1.5.5"
 nix = "0.25.0"
 rcgen = "0.10.0"
+docker-api = "0.11.0"
